Added support for pwm_configure_dev.

+++ hw/drivers/pwm/pwm_stm32/include/pwm_stm32/pwm_stm32.h
@@ -35,7 +35,7 @@ extern "C" {
 
 typedef struct stm32_pwm_conf {
     TIM_TypeDef   *tim;
-    uint32_t       irq;
+    uint16_t       irq;
 } stm32_pwm_conf_t;
 
 /**
